Add BeamInfo and DelayInfo support
Beams are autogenerated by `FringeExplorer.beamrings` function with ring
separation of 10 arcseconds.  This spacing should be configurable and a
custom set of beam positions should also be supported.  The delays and
delay rates are currently generated once per second for 300 seconds
starting from the start time of the RAW data.  Should probably use DWELL
to make sure we generate enough delays and rates.